# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rosana Frutos Secos — API
  version: 1.0.0
  description: 'Catálogo público y administración local de productos, categorías,
    galerías y stock. Precios en ARS como strings decimales; pesos enteros en gramos
    y stock en paquetes. Pedidos con reservas transaccionales, acceso por email y
    sesiones de operadores. El backoffice requiere operador activo, DEBUG, BACKOFFICE_ENABLED
    y cliente local; la documentación conserva su guard local. Escrituras requieren
    X-CSRFToken obtenido en /api/v1/auth/session/. Las operaciones de escritura de
    Swagger modifican datos reales: no ejecutarlas como prueba.'
paths:
  /api/v1/auth/customer/logout/:
    post:
      operationId: v1_auth_customer_logout_create
      summary: Cerrar acceso del cliente
      tags:
      - Autenticación
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
  /api/v1/auth/customer/request-link/:
    post:
      operationId: v1_auth_customer_request_link_create
      description: Respuesta genérica para evitar revelar cuentas. Enlace de un uso,
        enviado solo a emails con pedidos existentes. Requiere CSRF.
      summary: Solicitar acceso del cliente por email
      tags:
      - Autenticación
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailInputRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EmailInputRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EmailInputRequest'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
  /api/v1/auth/customer/verify/:
    post:
      operationId: v1_auth_customer_verify_create
      description: 'POST explícito: abrir un email no consume el enlace. Token aleatorio
        guardado como hash; no es el número de pedido. Requiere CSRF.'
      summary: Canjear enlace de acceso
      tags:
      - Autenticación
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyLinkRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VerifyLinkRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VerifyLinkRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
...
  /api/v1/auth/staff/login/:
    post:
      operationId: v1_auth_staff_login_create
      description: Solo operadores activos is_staff. Requiere X-CSRFToken incluso
        antes de iniciar sesión. Mantiene el guard de desarrollo local.
      summary: Iniciar sesión de operador
      tags:
      - Autenticación
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffLoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StaffLoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StaffLoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
  /api/v1/auth/staff/logout/:
    post:
      operationId: v1_auth_staff_logout_create
      summary: Cerrar la sesión compartida del navegador
      tags:
      - Autenticación
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
  /api/v1/backoffice/categories/:
    get:
      operationId: v1_backoffice_categories_list
      description: |+
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

      summary: Listar todas las categorías
      tags:
      - Backoffice
      security:
      - cookieAuth: []
      responses:
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminCategory'
          description: ''
    post:
      operationId: v1_backoffice_categories_create
      description: |+
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

      summary: Crear categoría
      tags:
      - Backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCategoryRequest'
            examples:
              Categoría:
                value:
                  name: Hierbas naturales
                  slug: hierbas-naturales
                  is_active: true
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminCategoryRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminCategoryRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: 'Validación: errores por campo, non_field_errors o lista de
            mensajes.'
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategory'
          description: ''
  /api/v1/backoffice/categories/{id}/:
    get:
      operationId: v1_backoffice_categories_retrieve
      description: |+
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

      summary: Consultar categoría
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este category.
        required: true
      tags:
      - Backoffice
      security:
      - cookieAuth: []
      responses:
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: El recurso no existe.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategory'
          description: ''
    patch:
      operationId: v1_backoffice_categories_partial_update
      description: |-
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

        No se puede desactivar si tiene productos publicados. No admite DELETE ni PUT.
      summary: Editar o desactivar categoría
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este category.
        required: true
      tags:
      - Backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAdminCategoryRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAdminCategoryRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAdminCategoryRequest'
      security:
      - cookieAuth: []
      responses:
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: 'Validación: errores por campo, non_field_errors o lista de
            mensajes.'
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: El recurso no existe.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategory'
          description: ''
  /api/v1/backoffice/images/{id}/:
    patch:
      operationId: v1_backoffice_images_partial_update
      description: |-
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

        Solo alt_text y credit; para reemplazar archivo o URL, agregar una imagen nueva.
      summary: Editar descripción o crédito de imagen
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este product image.
        required: true
      tags:
      - Backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedImageMetadataRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedImageMetadataRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedImageMetadataRequest'
      security:
      - cookieAuth: []
      responses:
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: 'Validación: errores por campo, non_field_errors o lista de
            mensajes.'
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: El recurso no existe.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductImage'
          description: ''
    delete:
      operationId: v1_backoffice_images_destroy
      description: |-
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

        Elimina la referencia, conserva el archivo físico para recuperación manual y normaliza posiciones. No se puede quitar la última imagen de un producto publicado: pasar antes a borrador.
      summary: Quitar imagen de la galería
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este product image.
        required: true
      tags:
      - Backoffice
      security:
      - cookieAuth: []
      responses:
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: 'Validación: errores por campo, non_field_errors o lista de
            mensajes.'
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: El recurso no existe.
        '204':
          description: No response body
  /api/v1/backoffice/orders/:
    get:
      operationId: v1_backoffice_orders_list
      description: Operador activo y guard local. Más antiguos primero; 50 por página,
        sin agenda.
      parameters:
      - in: query
        name: bucket
        schema:
          type: string
          enum:
          - cerrados
          - confirmar
          - entrega
          - entregados
          - proximos
        description: Columna del tablero; Próximos son reservados y preparando, sin
          agenda.
      - in: query
        name: delivery
        schema:
          type: string
          enum:
          - entrega_local
          - retiro_local
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
        description: Referencia o nombre, no modifica la identidad del cliente.
      - in: query
        name: status
        schema:
          type: string
          enum:
          - A_CONFIRMAR
          - CANCELADO
          - EN_REPARTO
          - LISTO_PARA_RETIRO
          - PREPARANDO
          - RESERVADO
          - TERMINADO
          - VENCIDO
      tags:
      - Backoffice
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminOrderList'
          description: ''
  /api/v1/backoffice/orders/{id}/:
    get:
      operationId: v1_backoffice_orders_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - Backoffice
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrder'
          description: ''
    patch:
      operationId: v1_backoffice_orders_partial_update
      description: Solo nota interna. No permite editar líneas, precios ni stock.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - Backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrderNoteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrderNoteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrderNoteRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrder'
          description: ''
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
  /api/v1/backoffice/orders/{id}/transition/:
    post:
      operationId: v1_backoffice_orders_transition_create
      description: Transacción con bloqueo de pedido, productos y variantes. expected_status
        evita acciones obsoletas. Reservar mantiene stock; cancelar/vencer libera;
        terminar consume físico y reservado con historial. Vencimiento automático
        solo en RESERVADO, no durante preparación/reparto.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f-]{36}$
        required: true
      tags:
      - Backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrder'
          description: ''
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
...
  /api/v1/backoffice/products/{id}/images/:
    post:
      operationId: v1_backoffice_products_images_create
      description: |-
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

        Un archivo image por solicitud multipart o image_url en JSON, nunca ambos. alt_text obligatorio. Máximo 10 imágenes por producto; archivos JPEG/PNG/WebP de hasta 5 MiB y 25 megapíxeles. Se agrega al final; posición 0 es portada.
      summary: Agregar imagen a la galería
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este product.
        required: true
      tags:
      - Backoffice
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageInputRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ImageInputRequest'
            examples:
              ImagenDesdeURL:
                value:
                  image_url: https://images.example.com/almendras.jpg
                  alt_text: Almendras naturales en un cuenco.
                  credit: ''
                summary: Imagen desde URL
        required: true
      security:
      - cookieAuth: []
      responses:
        '400':
          content:
            application/json:
              schema:
                oneOf:
                - type: object
                  additionalProperties: {}
                - type: array
                  items:
                    type: string
          description: 'Validación: errores por campo, non_field_errors o lista de
            mensajes.'
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: El recurso no existe.
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductImage'
          description: ''
...
  /api/v1/backoffice/variants/{id}/movements/:
    get:
      operationId: v1_backoffice_variants_movements_list
      description: |-
        Requiere DEBUG y BACKOFFICE_ENABLED, conexión loopback y Origin permitido si está presente. Requiere una sesión de operador activo y X-CSRFToken para escrituras.

        Hasta 100 movimientos, más recientes primero. Incluye delta, motivo, físico anterior/resultante y fecha.
      summary: Consultar historial de stock
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este product variant.
        required: true
      tags:
      - Backoffice
      security:
      - cookieAuth: []
      responses:
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: 'Acceso bloqueado: entorno, IP u origen no permitido.'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                required:
                - detail
          description: El recurso no existe.
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockMovement'
          description: ''
...
components:
  schemas:
...
    EmailInputRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
      required:
      - email
...
    Message:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
...
    PatchedImageMetadataRequest:
      type: object
      properties:
        alt_text:
          type: string
          minLength: 1
          maxLength: 255
        credit:
          type: string
          maxLength: 160
...
    Session:
      type: object
      properties:
        csrf_token:
          type: string
        staff:
          allOf:
          - $ref: '#/components/schemas/StaffIdentity'
          nullable: true
        customer:
          allOf:
          - $ref: '#/components/schemas/CustomerIdentity'
          nullable: true
      required:
      - csrf_token
      - customer
      - staff
...
    StaffLoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 256
      required:
      - password
      - username
...
    VerifyLinkRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 128
      required:
      - token
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
      description: Sesión HttpOnly. Escrituras requieren X-CSRFToken obtenido en /auth/session/.
...